retourne size_t ailieu d'un int

+++ ft_nbrlen.c
@@ -6,15 +6,15 @@
 /*   By: agaspar <agaspar@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/26 18:11:24 by agaspar           #+#    #+#             */
-/*   Updated: 2016/01/25 14:55:09 by agaspar          ###   ########.fr       */
+/*   Updated: 2016/01/25 15:00:13 by agaspar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_nbrlen(int n)
+size_t	ft_nbrlen(int n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	if (n < 0)
